Recipes/ENRT: split SRIOVDevices to standalone module
